requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.6
django-filter==2.4.0
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
types-python-dateutil==0.1.4
typing-extensions==3.10.0.0
